% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node.R
\name{add_node}
\alias{add_node}
\title{Add a new node to network}
\usage{
add_node(sbm, id, type = NULL, show_messages = TRUE)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network()}}.}

\item{id}{Unique identifying name for node.}

\item{type}{Type of node (string). This is used to distinguish multipartite
networks. E.g. "person", or "publication", etc..}

\item{show_messages}{Should function inform of its actions such as when a
model already exists so no changes are made?}
}
\value{
\code{sbm_network} object  with node added.
}
\description{
Add a node to the network. Takes the node id (string), the node type
(string), and the node level (int).
}
\examples{

my_sbm <- create_sbm() \%>\%
  add_node('node_1') \%>\%
  add_node('node_2')

get_state(my_sbm)

}
\seealso{
Other advanced: 
\code{\link{add_edge}()},
\code{\link{get_node_to_block_edge_counts}()},
\code{\link{initialize_blocks}()},
\code{\link{set_node_parent}()}
}
\concept{advanced}
