% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_run.R
\name{collapse_run}
\alias{collapse_run}
\title{Run Agglomerative Merging to targer a range of group numbers}
\usage{
collapse_run(
  sbm,
  level = 0,
  sigma = 2,
  start_group_num = 1,
  end_group_num = 10,
  num_group_proposals = 5,
  num_mcmc_sweeps = 10,
  beta = 1.5
)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{level}{Level of nodes who's groups will have their group membership run
through MCMC proposal-accept routine.}

\item{sigma}{Controls how fast collapse of network happens. For instance if
set to \code{2} then half (\code{1/2}) of the nodes will be removed at each step
until the desired number remains. If \code{exhaustive = TRUE} then this
parameter will do nothing}

\item{start_group_num}{Fewest number of groups to test}

\item{end_group_num}{Highest number of groups to test}

\item{num_group_proposals}{If \code{greedy = FALSE}, how many move proposals
should each node produce for merge options?}

\item{num_mcmc_sweeps}{Number of MCMC sweeps to run in between merge steps to
allow model to find best partitioning with given number of groups. Set to
zero for fast but more likely incorrect structure}

\item{beta}{Inverse temperature parameter for determining move acceptance
probability. Only applicable if \code{num_mcmc_sweeps > 0}.}
}
\value{
Tibble with three columns with rows corresponding to the result of
each merge step:  \code{entropy}, \code{num_groups} left in model, and a list column
of \code{state} which is the state dump dataframe for model at end of merge.
}
\description{
Run Agglomerative Merging to targer a range of group numbers
}
