% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combination_indices}
\alias{combination_indices}
\title{Get Combination Indices}
\usage{
combination_indices(n, repeats = FALSE)
}
\arguments{
\item{n}{Number of elements in vector to get combinations of}

\item{repeats}{Do we allow repeats in pairs? E.g. should a-a, b-b, etc. be allowed in generated pairs?}
}
\value{
A list with two vectors (\verb{$a} and \verb{$b}) containing integer indices of elements for every pair.
}
\description{
Generates a list with two arrays of indices that make up complete enumeration
of all combinations (order does \emph{not} matter) of elements of vector of length
\code{n}.
}
\examples{

my_letters <- head(letters, 5)

letter_pairs <- combination_indices(length(my_letters), repeats = FALSE)

dplyr::tibble(
  letter_1 = letters[letter_pairs$a],
  letter_2 = letters[letter_pairs$b]
)


}
\seealso{
Other helpers: 
\code{\link{build_score_fn}()},
\code{\link{get_sweep_pair_counts}()},
\code{\link{get_sweep_results}()},
\code{\link{print.sbm_network}()},
\code{\link{rolling_mean}()},
\code{\link{verify_model}()}
}
\concept{helpers}
