% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_class_method_wrappers.R
\name{compute_entropy}
\alias{compute_entropy}
\title{Compute Entropy}
\usage{
compute_entropy(sbm, level = 0)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{level}{Level of nodes for which group entropy is accessed. E.g. the
default value of \code{0} will get entropy of model as described by the groups
in level \code{1}.}
}
\value{
Entropy value (numeric).
}
\description{
Computes the (degree-corrected) entropy for the network at the specified
level (int).
}
