% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sbm.R
\name{create_sbm}
\alias{create_sbm}
\title{Create a new instance of an SBM model}
\usage{
create_sbm(edges = NULL, nodes = NULL, from_col = from, to_col = to)
}
\arguments{
\item{edges}{Dataframe with a from and two column encoding edges between
string node ids (direction does not matter). A list containing both a
\code{nodes} and \code{edges} dataframe can be passed to avoid having to unpack a
list of nodes and edges to the arguments of the same name. This is helpful
for the included simulation functions such as
\code{\link{sim_sbm_network}}.}

\item{nodes}{Optional dataframe that links a node \code{id} to its \code{type}, for when model is polypartite.}

\item{from_col}{Name of the from column for edges}

\item{to_col}{Name of the to column for edges}
}
\value{
Instance of SBM class.
}
\description{
Build an SBM instance from dataframe of edges. Optionally a nodes dataframe
can be provided to give information about node type.
}
\examples{
# Create an simple empty model
my_sbm <- create_sbm()

# Create model from simple simulated data
my_sbm <- create_sbm(sim_basic_block_network(n_blocks = 2, n_nodes_per_block = 25))


}
