% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_edge_counts.R
\name{get_block_edge_counts}
\alias{get_block_edge_counts}
\title{Get counts between blocks at a level}
\usage{
get_block_edge_counts(sbm, level = 1)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}
}
\value{
Dataframe of block pairs (\code{block_a} and \code{block_b}) and the number of edges between them (\code{count}).
}
\description{
Returns a dataframe with all pairs of blocks at the desired level and the number of edges between them. Unconnected blocks are ommitted for space savings.
}
\examples{
# Initialize a random network and assign random blocks
sbm <- sim_basic_block_network(n_blocks = 5) \%>\%
  create_sbm() \%>\%
  initialize_blocks(5)

# Gather edge counts for all level 1 block pairs
sbm \%>\% get_block_edge_counts(1)

}
