% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collapse_results.R
\name{get_collapse_results}
\alias{get_collapse_results}
\title{Get collapse results from model}
\usage{
get_collapse_results(sbm)
}
\arguments{
\item{sbm}{Object of class \code{sbm_network}.}
}
\value{
List with two dataframes. The first telling for all sweeps everytime
a node was moved and what group it was moved to. The second telling for
each sweep the entropy delta and total number of nodes that were moved to
new groups in that sweep. If \code{track_pairs = TRUE}, then an additional
\code{pairing_counts} dataframe is added to output.
}
\description{
Retrieves the dataframe recording the results of running either
\code{\link{collapse_blocks}} or \code{\link{collapse_run}} on
\code{sbm_network} object.
}
\examples{

# Start with a random network of two blocks with 25 nodes each and
# run agglomerative clustering with no intermediate MCMC steps on network
my_sbm <- sim_basic_block_network(n_blocks = 2, n_nodes_per_block = 25) \%>\%
  collapse_blocks(num_mcmc_sweeps = 0, sigma = 1.1)

# Look at the results of the collapse directly
get_collapse_results(my_sbm)

}
