% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entropy.R
\name{get_entropy}
\alias{get_entropy}
\title{Compute entropy for current model state}
\usage{
get_entropy(sbm)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network()}}.}
}
\value{
Entropy value (numeric).
}
\description{
Computes the (degree-corrected) entropy for the network at the node level.
}
\examples{

# Build basic network with 3 nodes and two blocks
my_sbm <- create_sbm() \%>\%
  add_node('node_1') \%>\%
  add_node('node_2') \%>\%
  add_node('node_3') \%>\%
  add_node('node_11', level = 1) \%>\%
  add_node('node_12', level = 1) \%>\%
  add_edge('node_1', 'node_2') \%>\%
  add_edge('node_1', 'node_3') \%>\%
  add_edge('node_2', 'node_3') \%>\%
  set_node_parent(child_id = 'node_1', parent_id = 'node_11') \%>\%
  set_node_parent(child_id = 'node_2', parent_id = 'node_11') \%>\%
  set_node_parent(child_id = 'node_3', parent_id = 'node_12')

# Compute entropy of network
get_entropy(my_sbm)

}
\seealso{
Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{get_block_edge_counts}()},
\code{\link{get_num_blocks}()},
\code{\link{get_state}()},
\code{\link{mcmc_sweep}()},
\code{\link{sample_posterior}()}
}
\concept{modeling}
