% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_edge_counts_at_level.R
\name{get_node_edge_counts_at_level}
\alias{get_node_edge_counts_at_level}
\title{Get node's edge counts to blocks}
\usage{
get_node_edge_counts_at_level(
  sbm,
  node_id,
  connection_level = 1L,
  node_level = 0L
)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{node_id}{String identifying the node that edge counts are desired for}

\item{connection_level}{Level of blocks to get connections to. E.g. \code{1} = node block level, \code{0} = connections to other nodes.}
}
\value{
Dataframe with two columns: \code{id} for the id of the node/block connected to, and \code{count} for the number of connections between.
}
\description{
For a given node in an SBM, returns a dataframe of each block (or node) that
the desired node is connected to and how many connections they have to that
block.
}
\examples{

# Initialize a random network and assign random blocks
sbm <- sim_basic_block_network(n_blocks = 5) \%>\%
create_sbm() \%>\%
initialize_blocks(5)

# Gather edge counts for a node in the network
sbm \%>\% get_node_edge_counts_at_level("g1_1")

}
