% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state.R
\name{get_state}
\alias{get_state}
\title{Dump state of current SBM model to dataframe}
\usage{
get_state(x)
}
\value{
A dataframe describing the current state of the network: \describe{
\item{id}{id of node} \item{parent}{id of node's parent. \code{none} if node has
no parent.} \item{type}{type of node} \item{level}{level of node in network
hierarchy} }
}
\description{
Exports the current state of the network as dataframe with each node as a row
and columns for node id, parent id, node type, and node level. Can also be
accomplished by \code{attr(sbm, 'state')}.
}
\examples{

# Build a bipartite network with 1 node and one block for each type
my_sbm <- create_sbm() \%>\%
  add_node('a1', type = 'a') \%>\%
  add_node('b1', type = 'b')

my_sbm \%>\% get_state()

}
