% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sweep_results.R
\name{get_sweep_results}
\alias{get_sweep_results}
\title{Get mcmc sweep results}
\usage{
get_sweep_results(sbm)
}
\arguments{
\item{sbm}{Object of class \code{sbm_network}.}
}
\value{
MCMC sweep results from an \code{sbm_network} object
}
\description{
Retrieves the MCMC sweep results from an \code{sbm_network} object. Needs to have
\code{\link{mcmc_sweep}} run on the object before.
}
\examples{

# Start with a small random network
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 15) \%>\%
  initialize_blocks(n_blocks = 4) \%>\%
  mcmc_sweep(num_sweeps = 4, variable_n_blocks = FALSE)

# Retrieve the sweep results from network
get_sweep_results(net)

}
\seealso{
Other helpers: 
\code{\link{build_score_fn}()},
\code{\link{get_combination_indices}()},
\code{\link{get_sweep_pair_counts}()},
\code{\link{print.sbm_network}()},
\code{\link{rolling_mean}()},
\code{\link{verify_model}()}
}
\concept{helpers}
