% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_model.R
\name{initialize_model}
\alias{initialize_model}
\title{Initialize model for sbm_network object}
\usage{
initialize_model(x, warn_if_overwriting = TRUE)
}
\arguments{
\item{x}{Object of class \code{sbm_network}.}

\item{show_messages}{Should function inform of its actions such as when a
model already exists so no changes are made?}
}
\value{

}
\description{
Takes data from sbm_network object and uses it to initialize an \code{SBM} class
model object.
}
\examples{
network <- sim_basic_block_network(
  n_blocks = 3,
  n_nodes_per_block = 40
)

sbm_net <- new_sbm_network(edges = network$edges, nodes = network$nodes)
sbm_net

sbm_net <- sbm_net \%>\% initialize_model()
sbm_net

}
