% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model__interblock_edge_counts.R
\name{interblock_edge_counts}
\alias{interblock_edge_counts}
\title{Get counts between blocks at a level}
\usage{
interblock_edge_counts(sbm, level = 1)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network}}.}

\item{level}{Level of nodes who's blocks will have their block membership run
through MCMC proposal-accept routine.}
}
\value{
Dataframe of block pairs (\code{block_a} and \code{block_b}) and the number of edges between them (\code{count}).
}
\description{
Returns a dataframe with all pairs of blocks at the desired level and the number of edges between them. Unconnected blocks are ommitted for space savings.
}
\examples{

# A small simulated network with random block assignment
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 10) \%>\%
  initialize_blocks(3)

# Get counts of connections from each given block to the others
net \%>\% interblock_edge_counts()

}
\seealso{
\code{\link{node_to_block_edge_counts}}

Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{entropy}()},
\code{\link{mcmc_sweep}()},
\code{\link{n_blocks}()},
\code{\link{state}()}
}
\concept{modeling}
