% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_class_method_wrappers.R
\name{load_from_state}
\alias{load_from_state}
\title{Load from state}
\usage{
load_from_state(sbm, state)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{state}{SBM state dump dataframe as provided by \code{\link{get_state}}.}
}
\value{
SBM model object modified with state modified to matched passed state
dump. \emph{Note that object is modified in place as well.}
}
\description{
Takes model state export as given by SBM$get_state and returns model to
specified state. This is useful for resetting model before running various
algorithms such as agglomerative merging.
}
