% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_load_sbm_network.R
\name{load_sbm_network}
\alias{load_sbm_network}
\title{Load sbm_network object}
\usage{
load_sbm_network(loc)
}
\arguments{
\item{loc}{Location on computer of saved model \code{.rds} file.}
}
\value{
New \code{sbm_network} object copy in same state as when model was saved.
}
\description{
Loads an SBM Network object that was saved by \code{\link{save_sbm_network}}
}
\examples{

network <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 40)

sbm_net <- new_sbm_network(edges = network$edges, nodes = network$nodes)

temp <- tempfile()
save_sbm_network(sbm_net, temp)

loaded_sbm_net <- load_sbm_network(tmp)

}
