% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_state.R
\name{load_state}
\alias{load_state}
\title{Load model state}
\usage{
load_state(sbm, state_dump)
}
\arguments{
\item{sbm}{SBM model object}

\item{state_dump}{Dataframe returned from \code{get_state(sbm)}}
}
\value{

}
\description{
Takes a state dump dataframe and updates the passed model to have that same structure.
}
