% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model__new_sbm_s4.R
\name{new_sbm_s4}
\alias{new_sbm_s4}
\title{Build new SBM s4 class}
\usage{
new_sbm_s4(
  nodes,
  edges,
  allowed_edge_types = NULL,
  node_types = NULL,
  random_seed = NULL,
  state_df = NULL
)
}
\arguments{
\item{nodes}{Optional dataframe that links a node \code{id} to its \code{type}, for
when model is polypartite.}

\item{edges}{Dataframe with a from and two column encoding edges between
string node ids (direction does not matter).}

\item{allowed_edge_types}{A dataframe in the same format of as \code{edges} that contains
allowed pairs (order does not matter) of possible node type combinations
across edges. For instance: nodes of type pollinator are allowed to connect
to flower nodes but not to other pollinator nodes. If this is left
undefined, it is inferred from edges.}

\item{node_types}{Character vector with unique node types in network. If not provided, this will be deduced from present types in \code{nodes}.}

\item{random_seed}{Integer seed to be passed to model's internal random
sampling engine. Note that if the model is restored from a saved state this
seed will be initialized again to the start value which will harm
reproducability.}

\item{state_df}{A state dataframe with \code{id}, \code{parent}, \code{level}, and \code{type} columns for all nodes in network (along with block nodes).}
}
\value{
An SBM S4 class
}
\description{
Builds an s4 class exported by Rcpp modules. Generally used internally. For making the main \code{sbm_network} class, use \code{\link{new_sbm_network}} instead.
}
\examples{

nodes <- dplyr::tribble(~id,     ~type,
"a1",    "a",
"a2",    "a",
"a3",    "a",
"b1",    "b",
"b2",    "b",
"b3",    "b",
"b4",    "b")

node_types <- c("a", "b")

edges <- dplyr::tribble(~a,     ~b,
                        "a1",    "b1",
                        "a1",    "b2",
                        "a1",    "b3",
                        "a2",    "b1",
                        "a2",    "b4",
                        "a3",    "b1")

random_seed <- 42
new_sbm_s4(nodes = nodes,
           edges = edges,
           node_types = node_types,
           random_seed = 42)

}
