% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model__node_to_block_edge_counts.R
\name{node_to_block_edge_counts}
\alias{node_to_block_edge_counts}
\title{Get node's edge counts to blocks}
\usage{
node_to_block_edge_counts(sbm, node_id, connection_level = 1L)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network}}.}

\item{node_id}{String identifying the node that edge counts are desired for}

\item{connection_level}{Level of blocks to get connections to. E.g. \code{1} = node block level, \code{0} = connections to other nodes.}
}
\value{
Dataframe with two columns: \code{id} for the id of the node/block connected to, and \code{count} for the number of connections between.
}
\description{
For a given node in an SBM, returns a dataframe of each block (or node) that
the desired node is connected to and how many connections they have to that
block.
}
\examples{

# A small simulated network with random block assignment
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 10) \%>\%
  initialize_blocks(3)

# Get a random node's edge counts to blocks
node_id <- sample(net$nodes$id, 1)
net \%>\% node_to_block_edge_counts(node_id, connection_level = 1)

}
\seealso{
\code{\link{interblock_edge_counts}}

Other advanced: 
\code{\link{add_edge}()},
\code{\link{add_node}()},
\code{\link{initialize_blocks}()},
\code{\link{update_state}()}
}
\concept{advanced}
