% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_collapse_results.R
\name{plot_collapse_results}
\alias{plot_collapse_results}
\title{Visualize agglomerative collapse results}
\usage{
plot_collapse_results(collapse_results, rolling_window = 3)
}
\arguments{
\item{collapse_results}{Dataframe of agglomerative merging based collapse
results as returned from \link{\code{collapse_run}}, or
\link{\code{collapse_groups(report_all_steps = TRUE)}}.}

\item{rolling_window}{Size of rolling window used to calculate each steps deviance from the rolling mean...}
}
\value{
GGplot object comparing the fit results and each step's deviance from the rolling mean
}
\description{
Visualize agglomerative collapse results
}
