% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmR-package.R
\docType{class}
\name{Rcpp_SBM-class}
\alias{Rcpp_SBM-class}
\alias{SBM}
\title{SBM C++ class}
\description{
This is the R reference to the c++ class that implements SBM functionality.
Almost all interfacing with this methods will be taken care of with the s3
class \code{sbm_network} and thus this should be ignored for the vast majority of
users. Documentation of all available methods is available in the c++ wrapper
code using the RCPP module protocol. See \code{\link[Rcpp]{loadModule}} for more info.
}
