% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample from SBM posterior}
\usage{
sample_posterior(
  edges,
  nodes = NULL,
  n_sweeps = 100,
  n_chains = 1,
  burnin_sweeps = 10,
  return_sweep_info = TRUE,
  starting_partition = "agglomerative",
  verbose = c("initialization"),
  ...
)
}
\arguments{
\item{edges}{Dataframe with a from and two column encoding edges between string node ids (direction does not matter).}

\item{nodes}{Optional dataframe that links a node \code{id} to its \code{type}, for when model is polypartite.}

\item{n_sweeps}{number of MCMC sweeps of all nodes (after burn-in)}

\item{n_chains}{How many seperate MCMC chains to start and run}

\item{burnin_sweeps}{How many sweeps to run before starting to record}

\item{return_sweep_info}{Return entropy and number of nodes changed for each
sweep (defaults to \code{TRUE})}

\item{starting_partition}{Algorithm to start each chain with? Current options
are \code{"agglomerative"} or \code{"random"}.}

\item{verbose}{Array of what info should be printed as model runs. Default
shows \code{"initialization"} info of model. Other options are \verb{"silent} to
print nothing, or \code{"iteration progress"} which prints out results of each
mcmc iteration as it happens.}

\item{...}{Additional arguments to \code{\link{create_sbm}}}
}
\value{

}
\description{
Sample from SBM posterior
}
