% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_load_sbm_network.R
\name{save_sbm_network}
\alias{save_sbm_network}
\title{Save sbm_network object}
\usage{
save_sbm_network(x, loc)
}
\arguments{
\item{x}{Object of class \code{sbm_network}.}

\item{loc}{Location on computer of saved model \code{.rds} file.}
}
\value{
Saved \code{.rds} file on disk at specified location.
}
\description{
Saves an SBM Network object to local disk for use accross sessions without instantiating again. Opened with \code{\link{load_sbm_network}}
}
\examples{

network <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 40)

sbm_net <- new_sbm_network(edges = network$edges, nodes = network$nodes)

temp <- tempfile()
save_sbm_network(sbm_net, temp)

loaded_sbm_net <- load_sbm_network(tmp)

}
\seealso{
Other model_setup: 
\code{\link{load_sbm_network}()},
\code{\link{new_sbm_network}()}
}
\concept{model_setup}
