% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_parent.R
\name{set_node_parent}
\alias{set_node_parent}
\title{Set the parent of a node}
\usage{
set_node_parent(sbm, child_id, parent_id, level = 0)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network}}.}

\item{child_id}{Id of node for which parent is being added}

\item{parent_id}{Id of parent node}

\item{level}{Level of child node. This will almost always stay at default
value of \code{0}.}
}
\value{
An S3 object of class \code{sbm_network}. For details see
\code{\link{new_sbm_network}} section "Class structure."
}
\description{
Sets the parent node (or block) for a given node. Takes child node's id
(string), parent node's id (string), and the level of child node (int).
}
\examples{

# Create a model with two nodes, one at the block level (1)
# Start with bipartite network with 6 nodes
net <- dplyr::tribble(
  ~a_node, ~b_node,
  "a1"   , "b1"   ,
  "a1"   , "b2"   ,
  "a1"   , "b3"   ,
  "a2"   , "b1"   ,
  "a3"   , "b1"
) \%>\%
  new_sbm_network(bipartite_edges = TRUE,
                  edges_from_col = a_node,
                  edges_to_col = b_node)

get_state(net)

# Make a parent node and assign it to a1
net <- set_node_parent(net, child_id = 'a1', parent_id = 'a1_parent')

get_state(net)

}
\seealso{
Other advanced: 
\code{\link{add_edge}()},
\code{\link{add_node}()},
\code{\link{get_node_to_block_edge_counts}()},
\code{\link{initialize_blocks}()},
\code{\link{update_state}()}
}
\concept{advanced}
