% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model__state.R
\name{state}
\alias{state}
\title{Dump state of current SBM model to dataframe}
\usage{
state(sbm)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network}}.}
}
\value{
A dataframe describing the current state of the network: \describe{
\item{id}{id of node} \item{parent}{id of node's parent. \code{none} if node has
no parent.} \item{type}{type of node} \item{level}{level of node in network
hierarchy} }
}
\description{
Exports the current state of the network as dataframe with each node as a row
and columns for node id, parent id, node type, and node level. Can also be
accomplished by \code{attr(sbm, 'state')}.
}
\examples{

# Small edge list to build network
edges <- dplyr::tribble(
  ~from, ~to,
  "a1"   , "b1"   ,
  "a1"   , "b2"   ,
  "a1"   , "b3"   ,
  "a2"   , "b1"   ,
  "a2"   , "b4"   ,
  "a3"   , "b1"
)

# A small simulated network with no blocks
net <- new_sbm_network(edges)

state(net)

# Add some blocks and state will reflect
net \%>\% initialize_blocks(2) \%>\% state()

}
\seealso{
Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{entropy}()},
\code{\link{interblock_edge_counts}()},
\code{\link{mcmc_sweep}()},
\code{\link{n_blocks}()}
}
\concept{modeling}
