% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_state.R
\name{update_state}
\alias{update_state}
\title{Update SBM model state}
\usage{
update_state(sbm, state_df)
}
\arguments{
\item{sbm}{Object of class \code{sbm_network}.}

\item{state_df}{A state dataframe with \code{id}, \code{parent}, \code{level}, and \code{type} columns for all nodes in network (along with block nodes).}
}
\value{
An S3 object of class \code{sbm_network}. For details see
\code{\link{new_sbm_network}} section "Class structure."
}
\description{
Updates the state of the internal S4 class along with the state holder of the s3 class.
}
\examples{

edges <- dplyr::tribble(
  ~from,   ~to,
  "a1",  "b1",
  "a1",  "b2",
  "a1",  "b3",
  "a2",  "b1",
  "a2",  "b4",
  "a3",  "b1"
)

nodes <- dplyr::tribble(
  ~id,   ~type,
  "a1", "node",
  "a2", "node",
  "a3", "node",
  "b1", "node",
  "b2", "node",
  "b3", "node",
  "b4", "node"
)

new_state <- dplyr::tribble(
  ~id,          ~parent,   ~type,  ~level,
  "a1",       "a_parent",  "node",      0L,
  "a2",       "a_parent",  "node",      0L,
  "a3",       "a_parent",  "node",      0L,
  "b1",       "b_parent",  "node",      0L,
  "b2",       "b_parent",  "node",      0L,
  "b3",       "b_parent",  "node",      0L,
  "b4",       "b_parent",  "node",      0L,
  "a_parent", "none",      "node",      1L,
  "b_parent", "none",      "node",      1L,
)

net <- new_sbm_network(edges = edges, nodes = nodes)

# before updating state the state should not equal the desired new state
dplyr::all_equal(new_state, attr(net, 'state'))
dplyr::all_equal(new_state, attr(net, 'model')$get_state())

# Update state using update_state() method
net <- net \%>\% update_state(new_state)

# Now the states should be identical
dplyr::all_equal(new_state, attr(net, 'state'))
dplyr::all_equal(new_state, attr(net, 'model')$get_state())

}
\seealso{
Other advanced: 
\code{\link{add_edge}()},
\code{\link{add_node}()},
\code{\link{get_node_to_block_edge_counts}()},
\code{\link{initialize_blocks}()},
\code{\link{set_node_parent}()}
}
\concept{advanced}
