% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_model.R
\name{verify_model}
\alias{verify_model}
\title{Verify model for sbm_network object exists}
\usage{
verify_model(x, show_messages = FALSE)
}
\arguments{
\item{x}{Object of class \code{sbm_network}.}

\item{show_messages}{Should function inform of its actions such as when a
model already exists so no changes are made?}
}
\value{

}
\description{
Takes data from sbm_network object and checks S4 \code{SBM} class model object
exists and is usable. If it isn't it is created.
}
\examples{
network <- sim_basic_block_network(
  n_blocks = 3,
  n_nodes_per_block = 40
)

sbm_net <- new_sbm_network(edges = network$edges, nodes = network$nodes)
sbm_net

sbm_net <- sbm_net \%>\% verify_model()
sbm_net

}
\seealso{
Other helpers: 
\code{\link{build_score_fn}()},
\code{\link{get_combination_indices}()},
\code{\link{print.sbm_network}()},
\code{\link{rolling_mean}()}
}
\concept{helpers}
