% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mcmc_trace.R
\name{visualize_mcmc_trace}
\alias{visualize_mcmc_trace}
\title{Trace plot of multiple MCMC sweep sweeps}
\usage{
visualize_mcmc_trace(sbm)
}
\arguments{
\item{sbm}{Object of class \code{sbm_network}.}
}
\value{
GGplot line plot with x-axis as sweep number and y-axes as sweep
entropy delta and number of nodes moved during sweep.
}
\description{
A simple line plot showing both the entropy delta and the number of nodes
moved for each sweep. \code{sbmR::mcmc_sweep()} must be run on object before this function.
}
\examples{

# Simulate a model and run mcmc sweeps on it
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 25) \%>\%
  initialize_blocks(3) \%>\%
  mcmc_sweep(num_sweeps = 75, track_pairs = TRUE)

# Plot results
visualize_mcmc_trace(net)

}
\seealso{
Other visualizations: 
\code{\link{visualize_collapse_results}()},
\code{\link{visualize_network}()},
\code{\link{visualize_propensity_dist}()},
\code{\link{visualize_propensity_network}()}
}
\concept{visualizations}
