% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize Network}
\usage{
visualize_network(edges, nodes = NULL, width = NULL, height = NULL)
}
\arguments{
\item{edges}{Dataframe with a from and two column encoding edges between
string node ids (direction does not matter). A list containing both a
\code{nodes} and \code{edges} dataframe can be passed to avoid having to unpack a
list of nodes and edges to the arguments of the same name. This is helpful
for the included simulation functions such as
\code{\link{sim_sbm_network}}.}

\item{nodes}{Optional dataframe that links a node \code{id} to its \code{type}, for when model is polypartite.}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Creates a simple interactive network visualization that visually encodes node membership and type (if present) for a passed network.
}
\examples{

set.seed(42)

# Simulate a 3 group/cluster network with 15 nodes per group
simulated_network <- sim_basic_block_network(n_groups = 3, n_nodes_per_group = 15)

# Visualize network
visualize_network(simulated_network)

}
