% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_propensity_dist.R
\name{visualize_propensity_dist}
\alias{visualize_propensity_dist}
\title{Visualize pairwise propensity distribution}
\usage{
visualize_propensity_dist(sbm, isolate_type = NULL)
}
\arguments{
\item{sbm}{Object of class \code{sbm_network}.}

\item{isolate_type}{Node type to be isolated for visualization. If left empty all types are included.}
}
\value{
GGplot2 plot containing ECDF of the pairwise propensities of residing in same block for MCMC sweeps.
}
\description{
Investigate the results of cluster stability from MCMC sweeps. Shows the
empirical CDF of the pairwise probability/proportion of two nodes residing in
the same block over the given MCMC sweeps.
}
\details{
\code{sbmr::mcmc_sweep(sbm, track_pairs = TRUE, ...)} must be run on object before this function.
}
\examples{

# Simulate a model
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 25) \%>\%
  initialize_blocks(3) \%>\%
  mcmc_sweep(num_sweeps = 75, track_pairs = TRUE)

# Plot results
visualize_propensity_dist(net)

}
\seealso{
Other visualizations: 
\code{\link{visualize_collapse_results}()},
\code{\link{visualize_mcmc_trace}()},
\code{\link{visualize_network}()},
\code{\link{visualize_propensity_network}()}
}
\concept{visualizations}
