% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_connection.R
\name{add_connection}
\alias{add_connection}
\title{Add connection/edge between two nodes in network}
\usage{
add_connection(sbm, a_node, b_node)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{a_node}{Id of first node in connection}

\item{b_node}{Id of second node in connection}
}
\value{
SBM model object modified with connection added. \emph{Note that object is modified in place as well.}
}
\description{
Connects two nodes in network (at level 0) by their ids (string).
}
\examples{

# Create SBM and add two nodes
my_sbm <- create_sbm() \%>\%
  add_node('node_1') \%>\%
  add_node('node_2')

# Add connection between two nodes
my_sbm \%>\% add_connection('node_1', 'node_2')

}
