% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edge.R
\name{add_edge}
\alias{add_edge}
\title{Add edge between two nodes in network}
\usage{
add_edge(
  sbm,
  from_node,
  to_node,
  from_node_type = NULL,
  to_node_type = NULL,
  show_messages = TRUE,
  type_warning = TRUE
)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network()}}.}

\item{from_node}{Id of first node in edge}

\item{to_node}{Id of second node in edge}

\item{from_node_type}{Type of \code{from_node}}

\item{to_node_type}{Type of \code{to_node}}

\item{show_messages}{Should function inform of its actions such as when a
model already exists so no changes are made?}

\item{type_warning}{If node types are not specified the model will issue a
warning and use the first node in networks type for both. To hide this
warning for unipartite networks set this to \code{TRUE}.}
}
\value{
An S3 object of class \code{sbm_network}. For details see
\link{new_sbm_network} section "Class structure."
}
\description{
Connects two nodes in network (at level 0) by their ids (string).
}
\examples{

# Start with network with 5 nodes all fully connected
net <- sim_random_network(n_nodes = 5, prob_of_edge = 1)

# Fully connected network will have 10 edges
attr(net, 'n_edges')

# Add a new edge between two non-yet-seen nodes
net <- net \%>\% add_edge('new_from_node', 'new_to_node')

# We have two nodes added to our model
attr(net, 'n_nodes')

# And one edge
attr(net, 'n_edges')

# adding an edge between two nodes already in the network
net <- net \%>\% add_edge('new_from_node', 'new_to_node')

# Another edge will be added
attr(net, 'n_edges')

}
\seealso{
Other advanced: 
\code{\link{add_node}()},
\code{\link{get_node_to_block_edge_counts}()},
\code{\link{initialize_blocks}()},
\code{\link{set_node_parent}()},
\code{\link{update_state}()}
}
\concept{advanced}
