% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node.R
\name{add_node}
\alias{add_node}
\title{Add a new node to network}
\usage{
add_node(sbm, id, type = "node", level = 0)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{id}{Unique identifying name for node.}

\item{type}{Type of node (string). This is used to distinguish multipartite networks. E.g. "person", or "publication", etc..}

\item{level}{Level in node hierarchy. Level \code{0} indicates node is at the data level, levels greater than \code{0} are group nodes.}
}
\value{
SBM model object modified with node added. \emph{Note that object is modified in place as well.}
}
\description{
Add a node to the network. Takes the node id (string), the node type
(string), and the node level (int).
}
\examples{

my_sbm <- create_sbm() \%>\%
  add_node('node_1') \%>\%
  add_node('node_2')

get_state(my_sbm)

}
