% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model__add_node.R
\name{add_node}
\alias{add_node}
\title{Add a new node to network}
\usage{
add_node(sbm, id, type = NULL, show_messages = TRUE)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network}}.}

\item{id}{Unique identifying name for node.}

\item{type}{Type of node (string). This is used to distinguish multipartite
networks. E.g. "person", or "publication", etc..}

\item{show_messages}{Should function inform of its actions such as when a
model already exists so no changes are made?}
}
\value{
An S3 object of class \code{sbm_network}. For details see
\code{\link{new_sbm_network}} section "Class structure."
}
\description{
Add a node to the network. Takes the node id (string), the node type
(string), and the node level (int).
}
\examples{

# Start with network with 5 nodes.
net <- sim_random_network(n_nodes = 5, prob_of_edge = 1, setup_model = TRUE)

# Verify nodes
net$nodes
net \%>\% state()

# Add two nodes
net <- net \%>\%
  add_node('new_node1', type = 'node') \%>\%
  add_node('new_node2', type = 'node')

# New nodes are added to node data and state
net$nodes
net \%>\% state()

}
\seealso{
Other advanced: 
\code{\link{add_edge}()},
\code{\link{initialize_blocks}()},
\code{\link{node_to_block_edge_counts}()},
\code{\link{update_state}()}
}
\concept{advanced}
