% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score.R
\name{calculate_collapse_score}
\alias{calculate_collapse_score}
\title{Calculate block collapse scores}
\usage{
calculate_collapse_score(
  sbm,
  heuristic,
  remove_state = TRUE,
  use_entropy = FALSE
)
}
\arguments{
\item{sbm}{Object of class \code{sbm_network}.}

\item{heuristic}{How the best partitioning is defined. Takes either a
function that takes one/two arguments: an entropy vector and an optional
number of blocks vector with each element corresponding to a given
location, or a string labeling algorithm. Currently only \code{"lowest"},
\code{"dev_from_rolling_mean"}, \code{"delta_ratio"}, \code{"trend_deviation"}, and \code{"nls_residual"} are supported.}

\item{remove_state}{Should the state column be removed from results? Will be
\code{TRUE} when visualizing results, \code{FALSE} when using score to choose best
state.}

\item{use_entropy}{Should the score be calculated with the entropy of the
collapse? If \code{FALSE}, the entropy delta will be used. The full entropy is
needed to make comparisons across separate collapses in the case of
\code{\link{collapse_run}}.}
}
\value{
Collapse results dataframe with a score column attached
}
\description{
A helper function that is used in \code{visualize_collapse_results} and
\code{choose_best_collapse_state}.
}
