% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_collapse_state.R
\name{choose_best_collapse_state}
\alias{choose_best_collapse_state}
\title{Choose and load best model state from agglomerative collapsing algorithm}
\usage{
choose_best_collapse_state(sbm, collapse_results, heuristic = "lowest")
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{collapse_results}{Results of running agglomerative collapse algorithm on sbm with \code{\link{collapse_groups}}.}

\item{heuristic}{How the best partitioning is defined. Currently only \code{'lowest' }is supported which chooses the point with lowest entropy.}
}
\value{
SBM model updated to match state from the merge step with 'best' partitioning.
}
\description{
Choose and load best model state from agglomerative collapsing algorithm
}
\examples{

}
