% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_class_method_wrappers.R
\name{collapse_groups}
\alias{collapse_groups}
\title{Collapse Groups}
\usage{
collapse_groups(sbm, level = 0, num_mcmc_sweeps = 10)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{level}{Level of nodes who's groups will have their group membership run
through MCMC proposal-accept routine.}

\item{num_mcmc_sweeps}{Number of MCMC sweeps to run in between merge steps to
allow model to find best partitioning with given number of groups. Set to
zero for fast but more likely incorrect structure}
}
\value{
List with \code{entropy} and model \code{state} after each merge.
}
\description{
Performs agglomerative merging on network, starting with each group has a
single node down to one group per node type. Arguments are level to perform
merge at (int) and number of MCMC steps to peform between each collapsing to
equilibriate group. Returns
}
