% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_entropy.R
\name{compute_entropy}
\alias{compute_entropy}
\title{Compute Entropy}
\usage{
compute_entropy(sbm, level = 0)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{level}{Level of nodes for which group entropy is accessed. E.g. the
default value of \code{0} will get entropy of model as described by the groups
in level \code{1}.}
}
\value{
Entropy value (numeric).
}
\description{
Computes the (degree-corrected) entropy for the network at the specified
level (int).
}
\examples{

# Build basic network with 3 nodes and two groups
my_sbm <- create_sbm() \%>\%
  add_node('node_1') \%>\%
  add_node('node_2') \%>\%
  add_node('node_3') \%>\%
  add_node('node_11', level = 1) \%>\%
  add_node('node_12', level = 1) \%>\%
  add_connection('node_1', 'node_2') \%>\%
  add_connection('node_1', 'node_3') \%>\%
  add_connection('node_2', 'node_3') \%>\%
  set_node_parent(child_id = 'node_1', parent_id = 'node_11') \%>\%
  set_node_parent(child_id = 'node_2', parent_id = 'node_11') \%>\%
  set_node_parent(child_id = 'node_3', parent_id = 'node_12')

# Compute entropy of network
compute_entropy(my_sbm)

}
