% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_sbm.R
\name{copy_sbm}
\alias{copy_sbm}
\title{Copy existing SBM object}
\usage{
copy_sbm(old_sbm, match_state = FALSE)
}
\arguments{
\item{old_sbm}{Old SBM created with \link{\code{create_sbm}}}

\item{match_state}{Should copy also match the current grouping state of old model?}
}
\value{
A new deep copy of the original SBM with parameters preserved.
}
\description{
Copy existing SBM object
}
