% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sbm.R
\name{create_sbm}
\alias{create_sbm}
\title{Create SBM}
\usage{
create_sbm(edges = NULL, nodes = NULL, from_col = from, to_col = to)
}
\arguments{
\item{edges}{Dataframe with a from and two column encoding edges between string node ids (direction does not matter).}

\item{nodes}{Optional dataframe that links a node \code{id} to its \code{type}, for when model is polypartite.}

\item{from_col}{Name of the from column for edges}

\item{to_col}{Name of the to column for edges}
}
\value{
Instance of SBM class.
}
\description{
Build an SBM instance from dataframe of edges. Optionally a nodes dataframe
can be provided to give information about node type.
}
