% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_blocks.R
\name{get_num_blocks}
\alias{get_num_blocks}
\title{Get number of blocks currently in model}
\usage{
get_num_blocks(sbm)
}
\value{
Number of clusters in current model state
}
\description{
Returns the number of blocks/clusters in the currenet model state.
}
\examples{
network <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 40, setup_model = TRUE) \%>\%
  initialize_blocks(num_blocks = 5)

# Returns the number of blocks currently in model. Note this is not neccesarily the true number.
get_num_blocks(network)

}
\seealso{
Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{get_block_edge_counts}()},
\code{\link{get_entropy}()},
\code{\link{get_state}()},
\code{\link{mcmc_sweep}()},
\code{\link{sample_posterior}()}
}
\concept{modeling}
