% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_blocks.R
\name{get_num_blocks}
\alias{get_num_blocks}
\title{Get number of blocks currently in model}
\usage{
get_num_blocks(sbm)
}
\value{
Number of clusters in current model state
}
\description{
Returns the number of blocks/clusters in the currenet model state.
}
\examples{

# Small edge list to build network
edges <- dplyr::tribble(
  ~from, ~to,
  "a1"   , "b1"   ,
  "a1"   , "b2"   ,
  "a1"   , "b3"   ,
  "a2"   , "b1"   ,
  "a2"   , "b4"   ,
  "a3"   , "b1"
)

# A small simulated network with no blocks
net <- new_sbm_network(edges)

get_num_blocks(net)

# Add some two random blocks
net \%>\% initialize_blocks(2) \%>\% get_num_blocks()

# Add a block per node
net \%>\% initialize_blocks() \%>\% get_num_blocks()

}
\seealso{
Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{get_block_edge_counts}()},
\code{\link{get_entropy}()},
\code{\link{get_state}()},
\code{\link{mcmc_sweep}()}
}
\concept{modeling}
