% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_class_method_wrappers.R
\name{get_state}
\alias{get_state}
\title{Get state}
\usage{
get_state(sbm)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}
}
\value{
A dataframe describing the current state of the network: \describe{
\item{id}{id of node}
\item{parent}{id of node's parent. \code{none} if node has no parent.}
\item{type}{type of node}
\item{level}{level of node in network hierarchy}
}
}
\description{
Exports the current state of the network as dataframe with each node as a row
and columns for node id, parent id, node type, and node level.
}
