% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_from_state.R
\name{load_from_state}
\alias{load_from_state}
\title{Load model state from state dump}
\usage{
load_from_state(sbm, state)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{state}{SBM state dump dataframe as provided by \code{\link{get_state}}.}
}
\value{
SBM model object modified with state modified to matched passed state
dump. \emph{Note that object is modified in place as well.}
}
\description{
Takes model state export as given by SBM$get_state and returns model to
specified state. This is useful for resetting model before running various
algorithms such as agglomerative merging.
}
\examples{

# A state dump dataframe as one would get from sbmR::get_state()
desired_state <- dplyr::tribble(
  ~id,  ~parent, ~type, ~level,
  "a1",   "a11",   "a",      0,
  "b1",   "b11",   "b",      0,
 "a11",  "none",   "a",      1,
 "b11",  "none",   "b",      1
)

# Create simple network with no block nodes
my_sbm <- create_sbm() \%>\%
  add_node('a1', type = 'a') \%>\%
  add_node('b1', type = 'b')

# Load our desired state
my_sbm \%>\% load_from_state(desired_state)
# Group nodes have been created and assigned to proper children nodes

}
