% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_class_method_wrappers.R
\name{mcmc_sweep}
\alias{mcmc_sweep}
\title{MCMC Sweep}
\usage{
mcmc_sweep(sbm, level = 0, variable_num_groups = TRUE)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{level}{Level of nodes who's groups will have their group membership run
through MCMC proposal-accept routine.}

\item{variable_num_groups}{Should the model allow new groups to be created or
empty groups removed while sweeping or should number of groups remain
constant?}
}
\value{
Total number of nodes that were changed in the previous sweep
}
\description{
Runs a single Metropolis Hastings MCMC sweep across all nodes at specified
level (For algorithm details see
\href{https://arxiv.org/abs/1705.10225}{Piexoto, 2018}). Each node is given a
chance to move groups or stay in current group and all nodes are processed in
random order. Takes the level that the sweep should take place on (int) and
if new groups groups can be proposed and empty groups removed (boolean).
}
