% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sweep.R
\name{mcmc_sweep}
\alias{mcmc_sweep}
\title{Run a single MCMC sweep over nodes}
\usage{
mcmc_sweep(sbm, level = 0, variable_num_blocks = TRUE)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{level}{Level of nodes who's blocks will have their block membership run
through MCMC proposal-accept routine.}

\item{variable_num_blocks}{Should the model allow new blocks to be created or
empty blocks removed while sweeping or should number of blocks remain
constant?}
}
\value{
List with array of all nodes that were moved in the
sweep (\code{nodes_moved}) and the sweeps impact on the model's entropy
(\code{entropy_delta}).
}
\description{
Runs a single Metropolis Hastings MCMC sweep across all nodes at specified
level (For algorithm details see
\href{https://arxiv.org/abs/1705.10225}{Piexoto, 2018}). Each node is given a
chance to move blocks or stay in current block and all nodes are processed in
random order. Takes the level that the sweep should take place on (int) and
if new blocks blocks can be proposed and empty blocks removed (boolean).
}
\examples{

set.seed(42)

# Start with a random network and assign randomly to 4 blocks
n_blocks <- 3
my_sbm <- create_sbm(sim_basic_block_network(n_blocks = n_blocks, n_nodes_per_block = 15)) \%>\%
  initialize_blocks(num_blocks = n_blocks)

# Calculate entropy with random blocks
compute_entropy(my_sbm)

# Run 4 MCMC sweeps
for(i in 1:4) my_sbm \%>\% mcmc_sweep(variable_num_blocks = FALSE)

# Calculate entropy after sweeps
compute_entropy(my_sbm)

}
