% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model__mcmc_sweep.R
\name{mcmc_sweep}
\alias{mcmc_sweep}
\title{Run a single MCMC sweep over nodes}
\usage{
mcmc_sweep(
  sbm,
  num_sweeps = 1,
  eps = 0.1,
  variable_num_blocks = TRUE,
  track_pairs = FALSE,
  level = 0,
  verbose = FALSE
)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network}}.}

\item{num_sweeps}{Number of times all nodes are passed through for move
proposals.}

\item{eps}{Controls randomness of move proposals. Effects both the block
merging and mcmc sweeps.}

\item{variable_num_blocks}{Should the model allow new blocks to be created or
empty blocks removed while sweeping or should number of blocks remain
constant?}

\item{track_pairs}{Return a dataframe with all pairs of nodes along with the
number of sweeps they shared the same group?}

\item{level}{Level of nodes who's blocks will have their block membership run
through MCMC proposal-accept routine.}

\item{verbose}{If set to \code{TRUE} then each proposed move for all sweeps will
have information given on entropy delta, probability of moving, and if the
move were accepted printed to the console.}
}
\value{
An S3 object of class \code{sbm_network}. For details see
\code{\link{new_sbm_network}} section "Class structure."
}
\description{
Runs a single Metropolis Hastings MCMC sweep across all nodes at specified
level (For algorithm details see
\href{https://arxiv.org/abs/1705.10225}{Piexoto, 2018}). Each node is given a
chance to move blocks or stay in current block and all nodes are processed in
random order. Takes the level that the sweep should take place on (int) and
if new blocks blocks can be proposed and empty blocks removed (boolean).
}
\examples{

set.seed(42)

# Start with a small simulated network with random block assignments
net <- sim_basic_block_network(n_blocks = 4, n_nodes_per_block = 15) \%>\%
  initialize_blocks(num_blocks = 4)

# Calculate entropy with random blocks
get_entropy(net)

# Run some MCMC sweeps
net <- mcmc_sweep(net, num_sweeps = 25, variable_num_blocks = FALSE)

# Entropy after sweeps
get_entropy(net)

# Per-sweep level information
get_sweep_results(net)

# Use track_pairs = TRUE to get an idea of node-pair similarity by looking at
# how often every pair of nodes is connected over sweeps
net \%>\%
  mcmc_sweep(num_sweeps = 25, track_pairs = TRUE) \%>\%
  get_sweep_pair_counts()


}
\seealso{
Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{get_block_edge_counts}()},
\code{\link{get_entropy}()},
\code{\link{get_state}()},
\code{\link{n_blocks}()}
}
\concept{modeling}
