% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample from model posterior}
\usage{
sample_posterior(
  edges,
  nodes = NULL,
  n_sweeps = 100,
  n_chains = 1,
  burnin_sweeps = 10,
  return_sweep_info = TRUE,
  starting_partition = "agglomerative",
  epsilon = 0.1,
  greedy = TRUE,
  N_CHECKS_PER_BLOCK = 5,
  initialization_mcmc_steps = 15,
  verbose = c("initialization"),
  ...
)
}
\arguments{
\item{...}{Additional arguments to \code{\link{new_sbm_network}}}
}
\value{

}
\description{
Function is currently under construction
}
\examples{

sample_posterior(sim_basic_block_network())

}
\seealso{
Other modeling: 
\code{\link{choose_best_collapse_state}()},
\code{\link{collapse_blocks}()},
\code{\link{collapse_run}()},
\code{\link{get_block_edge_counts}()},
\code{\link{get_entropy}()},
\code{\link{get_num_blocks}()},
\code{\link{get_state}()},
\code{\link{mcmc_sweep}()}
}
\concept{modeling}
