% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_posterior.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample from SBM posterior}
\usage{
sample_posterior(
  edges,
  nodes = NULL,
  n_sweeps = 100,
  n_chains = 1,
  burnin_sweeps = 10,
  return_sweep_info = TRUE,
  starting_partition = "agglomerative",
  beta = 1.5,
  epsilon = 0.1,
  greedy = TRUE,
  n_checks_per_group = 5,
  initialization_mcmc_steps = 15,
  verbose = c("initialization"),
  ...
)
}
\arguments{
\item{edges}{Dataframe with a from and two column encoding edges between string node ids (direction does not matter).}

\item{nodes}{Optional dataframe that links a node \code{id} to its \code{type}, for when model is polypartite.}

\item{n_sweeps}{Number of MCMC sweeps of all nodes (after burn-in)}

\item{n_chains}{How many seperate MCMC chains to start and run}

\item{burnin_sweeps}{How many sweeps to run before starting to record}

\item{return_sweep_info}{Return entropy and number of nodes changed for each
sweep (defaults to \code{TRUE})}

\item{starting_partition}{Algorithm to start each chain with? Current options
are \code{"agglomerative"} or \code{"random"}.}

\item{beta}{Inverse temperature parameter for MCMC acceptance probability}

\item{epsilon}{Parameter to inforce ergodicity of markov chain. Setting to
zero will remove ergodicity from model, setting high will make model moves
more random.}

\item{greedy}{If using agglomerative initialization should algorithm search
every possible merge?}

\item{n_checks_per_group}{If not using agglomerative initialization, how many
options for merging should be explored per group?}

\item{initialization_mcmc_steps}{If using agglomerative initialization, how
many steps of MCMC should be run after merging groups to allow model to
find optimal layout in new group partition?}

\item{verbose}{Array of what info should be printed as model runs. Default
shows \code{"initialization"} info of model. Other options are \verb{"silent} to
print nothing, or \code{"iteration progress"} which prints out results of each
mcmc iteration as it happens.}

\item{...}{Additional arguments to \code{\link{create_sbm}}}
}
\value{

}
\description{
Sample from SBM posterior
}
