% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmR-package.R
\docType{package}
\name{sbmr-package}
\alias{sbmr}
\alias{sbmr-package}
\title{Package reference}
\description{
Inference and visualization of network data using the Stochastic Block Model.
}
\details{
Imports
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tbilab.github.io/sbmr}
  \item \url{https://github.com/tbilab/sbmr}
}

}
\author{
\strong{Maintainer}: Nick Strayer \email{nick.strayer@gmail.com}

}
