% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmR-package.R
\docType{package}
\name{sbmR-package}
\alias{sbmR}
\alias{sbmR-package}
\title{sbmR}
\description{
Insert main package description here...
}
\details{
Imports
}
\author{
\strong{Maintainer}: Nick Strayer \email{nick.strayer@gmail.com}

}
