% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm_class_method_wrappers.R
\name{set_node_parent}
\alias{set_node_parent}
\title{Set node parent}
\usage{
set_node_parent(sbm, child_id, parent_id, level = 0)
}
\arguments{
\item{sbm}{SBM model object as created by \code{\link{create_sbm()}}.}

\item{child_id}{Id of node for which parent is being added}

\item{parent_id}{Id of parent node}

\item{level}{Level of child node. This will almost always stay at default
value of \code{0}.}
}
\value{
SBM model object modified with node hierarchy modified. Note that
object is modified in place as well.
}
\description{
Sets the parent node (or group) for a given node. Takes child node's id
(string), parent node's id (string), and the level of child node (int).
}
