% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_node_parent.R
\name{set_node_parent}
\alias{set_node_parent}
\title{Set the parent of a node}
\usage{
set_node_parent(sbm, child_id, parent_id, level = 0)
}
\arguments{
\item{sbm}{\code{sbm_network} object as created by
\code{\link{new_sbm_network()}}.}

\item{child_id}{Id of node for which parent is being added}

\item{parent_id}{Id of parent node}

\item{level}{Level of child node. This will almost always stay at default
value of \code{0}.}
}
\value{
\code{sbm_network} object
}
\description{
Sets the parent node (or block) for a given node. Takes child node's id
(string), parent node's id (string), and the level of child node (int).
}
\examples{

# Create a model with two nodes, one at the block level (1)
my_sbm <- create_sbm() \%>\%
  add_node('node_1') \%>\%
  add_node('node_11', level = 1)

# Assign node 11 as node 1's parent
my_sbm \%>\%
  set_node_parent(child_id = 'node_1', parent_id = 'node_11')

}
\seealso{
Other advanced: 
\code{\link{add_edge}()},
\code{\link{add_node}()}
}
\concept{advanced}
