% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_random_network.R
\name{sim_random_network}
\alias{sim_random_network}
\title{Simulate completely random network}
\usage{
sim_random_network(
  n_nodes = 5,
  prob_of_connection = 0.7,
  allow_self_connections = FALSE
)
}
\arguments{
\item{n_nodes}{Total number of unique nodes to simulate}

\item{prob_of_connection}{Probability of any given edges between two nodes
occuring}

\item{allow_self_connections}{Should nodes be allowed to have connections to
themselves?}
}
\value{
An edge dataframe with node ids for connections in the \code{from} and
\code{to} columns
}
\description{
Generate an edge dataframe from an \href{https://en.wikipedia.org/wiki/Erd\%C5\%91s\%E2\%80\%93R\%C3\%A9nyi_model}{Erdos-Renyi network.} An edge will occur
between two nodes with a uniform common probability.
}
\examples{
sim_random_network(n_nodes = 4, prob_of_connection = 0.9)
}
