% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sbm_network.R
\name{sim_sbm_network}
\alias{sim_sbm_network}
\title{Simulate network using stochastic block model}
\usage{
sim_sbm_network(
  block_info,
  edge_propensities,
  edge_dist = rpois,
  allow_self_edges = FALSE,
  keep_edge_counts = TRUE,
  setup_model = FALSE
)
}
\arguments{
\item{block_info}{A dataframe/tibble with two columns: \code{block}: the id of the
block, and \code{n_nodes}: the number of nodes to simulate from that block.}

\item{edge_propensities}{A dataframe with 3 columns: \code{block_1}: the id
of the from block, \code{block_2}: the id of the to block, and \code{propensity}: the
parameter for \code{edge_dist} that controls if and or how many edges should
occur between the pair of blocks on average.}

\item{edge_dist}{A distribution function that has two inputs: the first is
number of samples to draw and the second the \code{propensity} value from
\code{edge_propensities}. For instance, the default value is \code{rpois} which
takes two arguments: \code{n} and \code{lambda}. In this scenario the \code{propensity} of
edges between two nodes is provided to \code{lambda} and thus is the
average number of edges for each pair of nodes between two given
blocks.}

\item{allow_self_edges}{Should nodes be allowed to have edges to
themselves?}

\item{keep_edge_counts}{Should the edge counts stay on returned
edges? If edges distribution is a binary yes or no then you will likely
want to set this to \code{TRUE}.}

\item{setup_model}{Should an SBM model object be added? Set to \code{FALSE} if
network is just being visualized or described.}
}
\value{
An S3 object of class \code{sbm_network}. For details see
\link{new_sbm_network} section "Class structure."
}
\description{
Simulates nodes and edges from a network where edge edges are derived
from the Stochastic Block Model. A dataframe describing the different blocks
present and the number of nodes within them, a dataframe that provides the
\code{propensity} of edge between two edges, and a distribution function
who's main parameter the \code{propensity} value defines are needed.
}
\examples{
block_info <- dplyr::tribble(
  ~block, ~n_nodes,
     "a",       10,
     "b",       12,
     "c",       15
)

edge_propensities <- dplyr::tribble(
 ~block_1, ~block_2, ~propensity,
      "a",      "a",         0.7,
      "a",      "b",         0.2,
      "a",      "c",         0.1,
      "b",      "b",         0.9,
      "b",      "c",         0.4,
      "c",      "c",         0.4,
)

sim_sbm_network(block_info, edge_propensities, edge_dist = purrr::rbernoulli)

}
\seealso{
Other simulations: 
\code{\link{sim_basic_block_network}()},
\code{\link{sim_random_network}()}
}
\concept{simulations}
