% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_model.R
\name{verify_model}
\alias{verify_model}
\title{Verify model for sbm_network object exists}
\usage{
verify_model(x, show_messages = FALSE)
}
\arguments{
\item{x}{Object of class \code{sbm_network}.}

\item{show_messages}{Should function inform of its actions such as when a
model already exists so no changes are made?}
}
\value{

}
\description{
Takes data from sbm_network object and checks S4 \code{SBM} class model object
exists and is usable. If it isn't it is created.
}
\examples{

# Setup network without model initialized
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 20, setup_model = FALSE)

net

# Run verify model and the model should now be initialized
net <- net \%>\% verify_model()
net

}
\seealso{
Other helpers: 
\code{\link{build_score_fn}()},
\code{\link{get_combination_indices}()},
\code{\link{get_sweep_pair_counts}()},
\code{\link{get_sweep_results}()},
\code{\link{print.sbm_network}()},
\code{\link{rolling_mean}()}
}
\concept{helpers}
