% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_mcmc_sweep.R
\name{visualize_mcmc_trace}
\alias{visualize_mcmc_trace}
\title{Trace plot of multiple MCMC sweep sweeps}
\usage{
visualize_mcmc_trace(sbm)
}
\value{
GGplot line plot with x-axis as sweep number and y-axes as sweep
entropy delta and number of nodes moved during sweep.
}
\description{
A simple line plot showing both the entropy delta and the number of nodes
moved for each sweep. \code{sbmR::mcmc_sweep()} must be run on object before this function.
}
\examples{

set.seed(42)
# Start with a random network and assign randomly blocks
net <- sim_basic_block_network(n_blocks = 4, n_nodes_per_block = 15) \%>\%
  initialize_blocks(num_blocks = 5)

# Run 25 MCMC sweeps
sweep_results <- net \%>\% mcmc_sweep(num_sweeps = 25, variable_num_blocks = FALSE)

# Plot results
visualize_mcmc_trace(sweep_results)

}
\seealso{
Other visualizations: 
\code{\link{visualize_collapse_results}()},
\code{\link{visualize_network}()},
\code{\link{visualize_propensity_dist}()},
\code{\link{visualize_propensity_network}()}
}
\concept{visualizations}
