% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_network.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize network stucture}
\usage{
visualize_network(
  sbm,
  node_color_col = "block",
  node_shape_col = "type",
  width = "100\%",
  height = NULL
)
}
\arguments{
\item{node_color_col}{Name of column in the \code{nodes} dataframe that should be
encoded with color of nodes in the plot. If this value is numeric (defined
as more than 10 unique values that are numbers) the network will color
nodes on a linear scale from the min value to the highest value.}

\item{node_shape_col}{Name of column in the \code{nodes} dataframe that should be
encoded with shape of nodes in the plot}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Creates a simple interactive network visualization that visually encodes node
membership and type (if present) for a passed network.
}
\examples{

# Simulate a 3 block/cluster network with 15 nodes per block
net <- sim_basic_block_network(n_blocks = 3, n_nodes_per_block = 15)

# Visualize network with color encoding the blocks
visualize_network(net)

# Visualize network with shape encoding the blocks
visualize_network(net, node_shape_col = 'block', node_color_col = 'type')

}
\seealso{
Other visualizations: 
\code{\link{visualize_collapse_results}()},
\code{\link{visualize_mcmc_trace}()},
\code{\link{visualize_propensity_dist}()},
\code{\link{visualize_propensity_network}()}
}
\concept{visualizations}
