% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_propensity_network.R
\name{visualize_propensity_network}
\alias{visualize_propensity_network}
\title{Plot network of nodes connected by pairwise block propensity}
\usage{
visualize_propensity_network(sbm, proportion_threshold = 0.2)
}
\arguments{
\item{proportion_threshold}{Threshold of pairwise propensity to consider two nodes linked. Choose carefully!}
}
\value{
An interactive network plot with force layout.
}
\description{
Takes the results of \code{mcmc_sweep()} with \code{track_pairs = TRUE} and plots the
data as a network where two nodes have an edge between them if their pairwise
propensity to be in the same block is above some threshold.
}
\examples{

set.seed(42)
# Simulate network data and initialize model with it
net <- sim_basic_block_network(n_blocks = 3,n_nodes_per_block = 30)

# Run collapsing algorithm
collapse_results <- net \%>\% collapse_blocks(desired_num_blocks = 1, sigma = 1.1)

# Choose best collapse state
my_sbm <- choose_best_collapse_state(net, collapse_results, verbose = TRUE)

# Run MCMC sweeps and track pairs
sweep_results <- mcmc_sweep(my_sbm, num_sweeps = 100, eps = 0.4, track_pairs = TRUE)

# Plot connection propensity network
visualize_propensity_network(sweep_results, proportion_threshold = 0.4)

}
\seealso{
Other visualizations: 
\code{\link{visualize_collapse_results}()},
\code{\link{visualize_mcmc_trace}()},
\code{\link{visualize_network}()},
\code{\link{visualize_propensity_dist}()}
}
\concept{visualizations}
